All commands and events are now Addons
