--- BotCogs/__init__.py
@@ -1,8 +1,0 @@
-__all__ = [
-    "changelog",
-    "ping",
-    "play",
-    "reload",
-    "say",
-    "ver",
-]
