+++ logging.ini
@@ -17,10 +17,10 @@ handlers=consoleHandler, fileHandler
 qualname=BotCommands.play
 propagate=0
 
-[logger_importModules]
+[logger_importAddons]
 level=DEBUG
 handlers=consoleHandler, fileHandler
-qualname=import_all_modules
+qualname=func.import_addons
 propagate=0
 
 [handler_consoleHandler]
